# avoid double buffering + slight API clean up