--- a/.travis.yml
+++ b/.travis.yml
@@ -3,8 +3,7 @@ os:
   - linux
 if: branch = master OR tag IS present OR type = pull_request
 julia:
-  - 1.0
-  - 1.3
+  - 1.4
 notifications:
   email: false
 matrix:
@@ -14,7 +13,7 @@ codecov: true
 jobs:
   include:
     - stage: "Documentation"
-      julia: 1.0
+      julia: 1.4
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=".")); Pkg.instantiate()'