--- a/Project.toml
+++ b/Project.toml
@@ -7,7 +7,7 @@ version = "0.3.0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 
 [compat]
-julia = "1"
+julia = "1.4"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"