@jrevels 2020-05-01
Looks like the `read!` method I was using actually only handles views on Julia >= v1.4; I bumped the minimum Julia version out of laziness. I'm fine with dropping lower Julia versions if other folks are okay with that.